--- src/main/java/com/jp/backend/domain/comment/reposiroty/CommentRepository.java
@@ -1,5 +1,0 @@
-package com.jp.backend.domain.comment.reposiroty;
-
-public interface CommentRepository {
-
-}
