--- src/main/java/com/jp/backend/domain/comment/reposiroty/JpaReplyRepository.java
@@ -1,8 +1,0 @@
-package com.jp.backend.domain.comment.reposiroty;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.jp.backend.domain.comment.entity.Reply;
-
-public interface JpaReplyRepository extends JpaRepository<Reply,Long>,ReplyRepository {
-}
