--- src/main/java/com/jp/backend/domain/comment/reposiroty/ReplyRepository.java
@@ -1,4 +1,0 @@
-package com.jp.backend.domain.comment.reposiroty;
-
-public interface ReplyRepository {
-}
