+++ src/main/java/com/jp/backend/domain/comment/repository/CommentRepository.java
@@ -1,0 +1,5 @@
+package com.jp.backend.domain.comment.repository;
+
+public interface CommentRepository {
+
+}
