+++ src/main/java/com/jp/backend/domain/comment/repository/JpaReplyRepository.java
@@ -1,0 +1,8 @@
+package com.jp.backend.domain.comment.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.jp.backend.domain.comment.entity.Reply;
+
+public interface JpaReplyRepository extends JpaRepository<Reply, Long>, ReplyRepository {
+}
