+++ src/main/java/com/jp/backend/domain/comment/repository/ReplyRepository.java
@@ -1,0 +1,4 @@
+package com.jp.backend.domain.comment.repository;
+
+public interface ReplyRepository {
+}
